refactor: Use split_unfused as module name.
